+++ functions/ft_substr.c
@@ -6,15 +6,15 @@
 /*   By: azainabi <azainabi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/12 06:40:16 by azainabi          #+#    #+#             */
-/*   Updated: 2024/03/08 17:00:16 by azainabi         ###   ########.fr       */
+/*   Updated: 2024/04/23 04:00:27 by azainabi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minibash.h"
 
-static char	*cpy(unsigned int start, size_t len, char const *s, char *ptr)
+static char	*cpy(unsigned int start, int len, char const *s, char *ptr)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < len && s[start])
@@ -25,11 +25,11 @@ static char	*cpy(unsigned int start, size_t len, char const *s, char *ptr)
 	return (ptr);
 }
 
-char	*ft_substr(char const *s, unsigned int start, size_t len, t_all *all)
+char	*ft_substr(char const *s, unsigned int start, int len, t_all *all)
 {
 	char	*ptr;
-	size_t	i;
-	size_t	max_len;
+	int		i;
+	int		max_len;
 
 	i = 0;
 	if (!s)
@@ -37,7 +37,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len, t_all *all)
 	max_len = ft_strlen(s) - start;
 	if (len > max_len)
 		len = max_len;
-	if (start >= ft_strlen(s))
+	if (start >= (unsigned int)ft_strlen(s))
 	{
 		ptr = ft_malloc(1, 0, all);
 		if (!ptr)
